% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birth_dutch}
\alias{birth_dutch}
\title{Child birth data}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is numeric. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{birthwt}{The variable is numeric. Birthweight of the child}
}
}
\source{
\url{https://github.com/stefvanbuuren/brokenstick}
}
\usage{
birth_dutch
}
\description{
Measurements during ages 0-2 years for a representative sample of 1,933 Dutch children born in 1988-1989
}
\examples{
\dontrun{
birth_dutch
}
}
\keyword{datasets}
