% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birth_us}
\alias{birth_us}
\title{Child birth data}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is integer. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{birthwt}{The variable is integer. Birthweight of the child}
 \item{birthlen}{The variable is integer. Birth length of the child}
 \item{apgar1}{The variable is integer. a quick test performed on a baby at 1 minute after birth. https://medlineplus.gov/ency/article/003402.htm}
 \item{apgar5}{The variable is integer. a quick test performed on a baby at 5 minutes after birth. https://medlineplus.gov/ency/article/003402.htm}
 \item{mrace}{The variable is character. Race of the mother}
 \item{mage}{The variable is integer. Mother age at child birth}
 \item{smoked}{The variable is integer. If the mother smoked (1) or didn't smoke (0)}
 \item{meducyrs}{The variable is integer. Educational years of mother}
 \item{ses}{The variable is character. Socioeconomic status of mother}
}
}
\source{
\url{https://github.com/hafen/hbgd}
}
\usage{
birth_us
}
\description{
Subset of growth data from the collaborative perinatal project (CPP).
}
\examples{
\dontrun{
birth_us
}
}
\keyword{datasets}
