% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childhealth_dutch}
\alias{childhealth_dutch}
\title{Child height and weight measurements}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is numeric. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{agedays}{The variable is numeric. Age in days}
 \item{gagebrth}{The variable is numeric. Gestational age at birth (days)}
 \item{htcm}{The variable is numeric. Length/height in cm (34-102)}
 \item{wtkg}{The variable is numeric. Weight measurement in kg (0.8-20.5)}
 \item{haz}{The variable is numeric. Height in SDS relative to WHO child growth standard}
 \item{waz}{The variable is numeric. Weight in SDS relative to WHO child growth standard}
}
}
\source{
\url{https://github.com/stefvanbuuren/brokenstick}
}
\usage{
childhealth_dutch
}
\description{
Longitudinal height and weight measurements during ages 0-2 years for a representative sample of 1,933 Dutch children born in 1988-1989.
}
\examples{
\dontrun{
childhealth_dutch
}
}
\keyword{datasets}
