% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childhealth_maled}
\alias{childhealth_maled}
\title{Child height, weight, head circumference measurements}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is character. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{country}{The variable is character. Label for the varied countries}
 \item{agedays}{The variable is numeric. Age in days}
 \item{wtkg}{The variable is numeric. Weight measurement in kg (0.8-20.5)}
 \item{stcm}{The variable is numeric. Stature either Length or height in cm}
 \item{htcm}{The variable is numeric. Height in cm}
 \item{lncm}{The variable is numeric. Length in cm}
 \item{lh_used}{The variable is character. Lenght or Height used for stature}
 \item{hccm}{The variable is numeric. Head Circumference in cm}
 \item{lhaz}{The variable is numeric. Length or Height for age in SDS relative to WHO child growth standard}
 \item{haz}{The variable is numeric. Height for age in SDS relative to WHO child growth standard}
 \item{laz}{The variable is numeric. Length for age in SDS relative to WHO child growth standard}
 \item{waz}{The variable is numeric. Weight for age in SDS relative to WHO child growth standard}
 \item{hcaz}{The variable is numeric. Head cirumference for age in SDS relative to WHO child growth standard}
 \item{whz}{The variable is numeric. Weight for height or lenght in SDS relative to WHO child growth standard}
}
}
\source{
\url{https://www.fic.nih.gov/About/Staff/Pages/mal-ed.aspx and https://clinepidb.org/ce/app/record/dataset/DS_5c41b87221}
}
\usage{
childhealth_maled
}
\description{
Subset of growth data from the Malnutrition and Enteric Disease Study (MAL-ED).
}
\examples{
\dontrun{
childhealth_maled
}
}
\keyword{datasets}
