% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childhealth_summary}
\alias{childhealth_summary}
\title{Child height and weight HAZ summaries}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is character. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{country}{The variable is character. Label for the varied countries}
 \item{haz_mean}{The variable is numeric. The average HAZ score over all measurements}
 \item{waz_mean}{The variable is numeric. The average WAZ score over all measurements}
 \item{observations}{The variable is integer. Number of observations for that subject}
 \item{agedays_last}{The variable is numeric. The age in days for the HAZ and WAZ last variable}
 \item{agedays_first}{The variable is numeric. The age in days for the HAZ and WAZ first variable}
 \item{haz_last}{The variable is numeric. The first HAZ measurement on the subject for age in days}
 \item{haz_first}{The variable is numeric. The last HAZ measurement on the subject for age in days}
 \item{waz_first}{The variable is numeric. The first WAZ measurement on the subject for age in days}
 \item{waz_last}{The variable is numeric. The last WAZ measurement on the subject for age in days}
}
}
\source{
\url{https://github.com/stefvanbuuren/brokenstick, https://github.com/hafen/hbgd, and https://clinepidb.org/ce/app/record/dataset/DS_5c41b87221}
}
\usage{
childhealth_summary
}
\description{
For all data from three studies
}
\examples{
\dontrun{
childhealth_summary
}
}
\keyword{datasets}
