% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{childhealth_us}
\alias{childhealth_us}
\title{Child height and weight measurements}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is integer. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{agedays}{The variable is integer. Age in days}
 \item{gagebrth}{The variable is numeric. Gestational age at birth (days)}
 \item{htcm}{The variable is integer. Length/height in cm (34-102)}
 \item{wtkg}{The variable is numeric. Weight measurement in kg (0.8-20.5)}
 \item{haz}{The variable is numeric. Height in SDS relative to WHO child growth standard}
 \item{waz}{The variable is numeric. Weight in SDS relative to WHO child growth standard}
 \item{mrace}{The variable is character. Race of the mother}
 \item{mage}{The variable is integer. Mother age at child birth}
 \item{meducyrs}{The variable is integer. Educational years of mother}
 \item{ses}{The variable is character. Socioeconomic status of mother}
}
}
\source{
\url{https://github.com/hafen/hbgd}
}
\usage{
childhealth_us
}
\description{
Subset of growth data from the collaborative perinatal project (CPP).
}
\examples{
\dontrun{
childhealth_us
}
}
\keyword{datasets}
