% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{days_365}
\alias{days_365}
\title{Child height and weight measurements for all data from three studies at one year of age.}
\format{
A data frame with columns:
\describe{
 \item{subjid}{The variable is character. unique identifyer of each child}
 \item{sex}{The variable is character. Male or Female}
 \item{htcm}{The variable is numeric. Height in cm}
 \item{wtkg}{The variable is numeric. Weight measurement in kg (0.8-20.5)}
 \item{haz}{The variable is numeric. Height for age in SDS relative to WHO child growth standard}
 \item{waz}{The variable is numeric. Weight for age in SDS relative to WHO child growth standard}
 \item{country}{The variable is character. Label for the varied countries}
}
}
\source{
\url{https://github.com/stefvanbuuren/brokenstick, https://github.com/hafen/hbgd, and https://clinepidb.org/ce/app/record/dataset/DS_5c41b87221}
}
\usage{
days_365
}
\description{
A subset of measurements from the three studies.
}
\examples{
\dontrun{
days_365
}
}
\keyword{datasets}
