% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weight_coef}
\alias{weight_coef}
\title{WHO coeficients for weight Z-score calculations}
\format{
A data frame with columns:
\describe{
 \item{agedays}{The variable is numeric. Age of child in days}
 \item{sex}{The variable is character. The gender of the child}
 \item{mean}{The variable is numeric. The mean for a child at agedays}
 \item{sd}{The variable is numeric. The standard deviation of weights for children at agedays}
 \item{cv}{The variable is numeric. The coefficient of variation for children at agedays - https://en.wikipedia.org/wiki/Coefficient_of_variation}
 \item{l}{The variable is numeric. The box-cox tranformation for children at agedays - https://www.cdc.gov/nchs/data/nhsr/nhsr063.pdf}
}
}
\source{
\url{https://github.com/hafen/growthstandards}
}
\usage{
weight_coef
}
\description{
See https://www.cdc.gov/nchs/data/nhsr/nhsr063.pdf for a desciption on how calculations are made.  However, the CDC has different coefficients.
}
\examples{
\dontrun{
weight_coef
}
}
\keyword{datasets}
